LiveCursors: recompute positions on scroll
Since everything is viewport-relative we need to recompute on scroll.
Maybe it shouldn't be viewport-relative, but this change was the path of
least resistance (if ugly) and we can improve later.

(Moving to page-relative is hard to actually render, since the component
can be anywhere on the page. I guess we could insert a div into the
`<body>` and portal the cursors into there, but that's also not super
pretty and has BC issues too.)


Test Plan:
Put a few cursors on the screen, they recompute position on scroll now.

Reviewers: AlbertQM, flooey, lazybean

Reviewed By: AlbertQM

Pull Request: https://github.com/getcord/monorepo/pull/6551

monorepo-commit: 9cd1cf7115e32d6a0e1add7dd38a69345b53a74d