LiveCursors: export LiveCursorsDefaultCursor
I exported the props but not the element itself. This could be useful to
folks customising LiveCursors. It's not really a useful component on its
own worth documenting, but it's useful enough to export for folks
looking for it.


Test Plan: Everything builds.

Reviewers: flooey

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/6570

monorepo-commit: f54b821731ee7d2ded54f8ead5184a077e05ff1a