Add file upload REST API (#6467)
Add an API for uploading files.  This creates a file and uploads it in
one call, and then returns the file ID, which can be used in message
attachments.

One downside here is that the owner has to be encoded as
multipart/form-data rather than JSON.  This is generally easy enough
with various tools (fetch, axios, and curl all provide built-in
functionality for encoding as multipart/form-data), but it's
inconsistent with our other APIs.  I think that's okay here, but
wanted to call it out.


Test Plan:

Added a file over the API, and then created a message with that file
attached:

```
$ ./scripts/testbedcurl.sh https://local.cord.com:8161/v1/files -F 'file=@/Users/flooey/Downloads/valves.png' -F ownerID=flooey
{
  "success": true,
  "message": "File created.",
  "fileID": "ac7fe13f-788e-4b4b-aec3-dac7b713ef64"
}

$ ./scripts/testbedcurl.sh https://local.cord.com:8161/v1/threads/abc123/messages -X POST --json '{"content": [], "addAttachments": [{"type": "file", "id": "ac7fe13f-788e-4b4b-aec3-dac7b713ef64"}], "authorID": "flooey"}'
{
  "success": true,
  "message": "Message created.",
  "messageID": "cord:7534f417-8f71-4353-b780-36d1a9a474f6"
}
```

The message appeared with the file attached.

Did the same with a custom S3 bucket set for the application, the file
ended up in the proper bucket.

monorepo-commit: 6bde95feae18b773822fead45965632648550ea0