Enable i18n in open source components (#6536)
Make it so that open source components (like ThreadedComments) can use
the i18n system.

This involves a few things.  The definition of the strings (which is
needed for proper typing) is moved to `@cord-sdk/types`.  Instead of
using the default i18n instance, create an explicit instance, expose
it from the CordSDK object, and provide it explicit in the root React
stack and CordProvider.

As a proof-of-concept, I've translated two strings in ThreadedComments
(the headers for the open and resolved tabs).  There are lots of other
strings to be extracted, but that will be done in future diffs.

I considered an alternative of having a separate i18n instance for the
open source code, instead of sharing it between them, but that would
mean that developers have to know whether the component they're using
is open source or not in order to know where to set their
translations, which is really confusing.

One downside here is that we can no longer render ThreadedComments (or
any open source component that contains text) before the SDK is
loaded, because the strings are supplied by the SDK.  That seems okay,
since it doesn't really do much at that point (none of the components
can render or have data before the SDK is loaded).

Fixes E-2917

Test Plan:

Loaded ThreadedComments, it had proper translations from first render.
Updated the translations via `updateOptions` and they changed in the
component.

monorepo-commit: c51c6e1db3e1506d79ac172a54a15929750e5159