Add cord-threaded-comments to the component list (#6601)
We missed out on adding the cord-threaded-comments webcomponent to the
list of web components, so the JSX package didn't tell JSX that it was
a valid tag name, meaning using it would fail typechecking.

Test Plan:

Added a `cord-threaded-comments` component to a test page, before this
diff it had an error about an unknown JSX element, now it does not.

monorepo-commit: c4e3be7fbe7d3f5e326f51a5d0c025ae4582d817