Fixing ThreadedComments docs 500 error on refresh
When navigating to the ThreadedComments docs via another page of the docs, it
loaded fine. But when refreshing it, we got a 500 error with a log saying: 'ReferenceError: window is not defined'.

This is because our docs are server-side rendered and the window object isn't available. I have added another check
to make sure that we don't try an access the object before it's defined.

Test Plan:
Refreshing the ThreadedComments docs is fine!


Reviewers: flooey

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/6626

monorepo-commit: 8d2620e5c5ec5347e31b32260e0520a94df2d2be