Adding threadMetadata to the ThreadedComments component
Creating a prop on the Composer and ThreadedComments components to accept thread metadata. The Composer can already accept message metadata, but for the ThreadedComments use case, adding message metadata doesn't really make sense, as the filtering would work on the thread level.

The prop behaves the same way as the threadName and threadUrl, meaning that the metadata can only be set when creating a thread and cannot be changed through the component.

Test Plan:
Added threadMetadata on the testbed threaded comments. Saw the metadata correctly in the db.


Reviewers: jwatzman, flooey

Reviewed By: jwatzman

Pull Request: https://github.com/getcord/monorepo/pull/6603

monorepo-commit: d1969a3853ec05bcec6bdaa980aa7386ab1e4243