ThreadedComments: filter prop
Adding a filter prop. Just passing it down into the LocationData API through the ThreadedComments props. There is a bit of tension with the `displayResolved` prop which already filters using this parameter. We decided to make sure that the `displayResolved` prop always takes precedence over the resolved status, if both are specified. This may be a bit confusing if developers haven't read the docs properly, but it will ensure that no weird configuration is done between the props and the component behaves as expected.

Test Plan:
Filtered by metadata and saw the correct threads in ThreadedComments in the testbed. Also tested filtering for resolved while setting displayResolved to unresolvedOnly. The displayResolved prop took precedence.


Reviewers: jwatzman, flooey

Reviewed By: jwatzman

Pull Request: https://github.com/getcord/monorepo/pull/6611

monorepo-commit: 6295d707363d3e04e8b2021c734dec56b05717bf