Standardize i18n user interpolation (#6627)
Make there be a standard object that is sent to the i18n system when
users are involved in interpolation.  In this case, it's
ClientUserData (the object generated in the client user API) plus the
`displayName` and `fullName` fields we use for standard display.

Clean up all the places where we provide users to translation strings,
which in some cases means passing through additional information to
the components.  In general, this doesn't require any new data, just
changes in types, except for the mention list, which previously didn't
have metadata or a couple underlying data fields.

Fixes E-2883

Test Plan:

Changed the translation string for viewer_user to be
"{{user.displayName}} {{user.metadata.pronouns}} (you)" with a user
that had that metadata set, it worked.

monorepo-commit: fcad97e809fe614b0bd773d7ac288452af31ea56