Standardize i18n keys (#6644)
Add a few patterns to the i18n keys that make them more regular.  This
should make it both easier for developers to understand where they are
used and make it easier to name future strings.

I didn't attempt to update any of the `annotation` strings, and the
`threaded_comments` strings can't be changed because they are
referenced in open source code, and thus might be referenced under the
old names still.  If we want to rename the latter, we can leave a
transitional key in, but I didn't think it was worth it right now.

Test Plan:

Everything compiles

monorepo-commit: d98005094d238d34babbf2c49ff3f807d4202f18