ThreadedComments: Don't flash "Load more" on render
We were simply using the API incorrectly: We need to check we're not `loading`, because the `hasMore` value jumps from `true` to `false` (and the former "Fetch more" would flash on screen).

Test Plan:
Add an empty ThreadedComments to the testbed.
Load the page, the "Load more" doesn't flash.
Add a ton of comment, refresh the page, check the "Load more" button is there and works correctly.


Reviewers: flooey, nickfil22

Reviewed By: nickfil22

Pull Request: https://github.com/getcord/monorepo/pull/6670

monorepo-commit: 3d9bb99dbbf305617bab3a35c2b820cc0890fbfb