Privacy: don't require organization_id in `getClientAuthToken`
This requirement was removed from the TS type a while ago, let's remove
it from the runtime check as well.

An org ID is still required right now outside of a couple of internal
experiments. However, the purpose of the check here is to prevent
accidentally minting a server auth token when you intended to mint a
client auth token, and checking specifically for `user_id` (which is not
optional and we do not intend to make optional) is good enough for that.
When the token is used, the server will do the correct complex check and
complain if an org ID is missing when it should be present.


Reviewers: flooey, yeo-yeo

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/6690

monorepo-commit: a9cf9904706174cd297c8fcf95be6baa0ad2725a