Extract the strings in ThreadedComments (#6715)
Extract all the strings in ThreadedComments into translatable strings.

Because the translations depend on the CordSDK being loaded, I had to
move the fetch more button into its own component so it wouldn't try
to render before the SDK was loaded, otherwise it explodes because it
violates the rules of hooks.

I tried to make it so the translations matched other components, but
mostly the strings are different.  The only ones that match up are the
resolved status and unresolve action, which I made reference the other
strings.

Part of E-2879

Test Plan:

Rendered a ThreadedComments and changed language to `cimode`, no
English strings showed up except the message text.

Rendered a ThreadedComments and updated some of the translations via
`cord.updateOptions()`, they updated in the UI.

monorepo-commit: e464660f041172ed75a40496007b608b89746c0b