onThreadResolve and onThreadReopen events on the message component
A message can be resolved and reopened via the optionsMenu, if it's the first message. This isn't great, but the alternative of actually creating an optionsMenu component is not the best thing to do right now. We need these events for ThreadedComments so we will be adding the events here, at least temporarily.

Because this is the case - is it even worth documenting on the message component? I think it might not be? Thoughts?

I decided to use the ThreadListContext to pass the events down, as the other option of prop drilling had a significant number of layers to go through.

Test Plan:
Passed callback functions for onThreadResolve and onThreadReopen. The correct data was console.log-ged.
Also tested in #6533


Reviewers: dmmiller, flooey, jwatzman

Reviewed By: jwatzman

Pull Request: https://github.com/getcord/monorepo/pull/6649

monorepo-commit: d2773039d71f257d39b2326548778d2e88241cd5