onThreadReopen event on the Composer component
If we set the thread-id to a composer component, there is a possibility this thread will be resolved. When this is the case, the composer is in the state shown in the video below. Because we can reopen the thread when clicking the button, I have added a `threadreopen` event to the component. Eventually, we want to use this in ThreadedComments, but we also need to have it here for it to work.

Test Plan:
Added a composer to the testbed and console.log-ged on thread reopen, which worked fine. Then added a web-component and an event listener, the event fired when reopening!
Also tested in #6533


https://github.com/getcord/monorepo/assets/102963329/9c8e4906-099b-45a9-928e-24120ec32183




Reviewers: elenaperezrioja, flooey, dmmiller, jwatzman

Reviewed By: jwatzman

Pull Request: https://github.com/getcord/monorepo/pull/6651

monorepo-commit: 8ba55f5fc74a49e8bf1c61aa534449a228e1629b