ThreadedComments: onThreadResovle/onThreadReopen events
Adding these two events to the ThreadedComments functionality. We are basically writing code around our `useThreadSummary` API, to detect changes in the resolved status of threads and exposing that as an offering
of the ThreadedComments component, even though technically, developers could do that in the exact same
way.

We are passing the events through to the Message and Composer components, which will fire the relevant events
from the optionsMenu or resolvedComposer.

Test Plan:
https://github.com/getcord/monorepo/assets/102963329/caad56a4-0370-4fb1-b106-94cc32040b1d


https://github.com/getcord/monorepo/assets/102963329/97cca97f-dc4d-44d4-bd3a-cb25a1439f3b




Reviewers: jwatzman, flooey

Reviewed By: jwatzman

Pull Request: https://github.com/getcord/monorepo/pull/6533

monorepo-commit: 5deec00fe12f0a25359a0525cb12610f6f126989