Rename observeThread summary to thread (#6774)
The field named `summary` in the return value of `observeThread`
didn't sit right to me, because it contains more than just the
summary, it also includes the ID, URL, location, subscribers,
metadata, etc.  Rename it to `thread`, which better describes the
contents.  Leave a getter in place for `summary` which logs the call,
so we can see if we can eventually delete it.

Also clean up a little bit of language in the description of the
fields on `ClientThreadData`, and use `null` rather than `undefined`
as the value for missing, which is more consistent with our other
APIs.

Also make the React hook just return the value from the callback
rather than splitting it apart and reassembling it, it's much simpler
that way.

Test Plan:

Everything still compiles.

Called observeThread(), I got back the same value for both fields.

monorepo-commit: e5c75b389aaece865aa9009a22f09ec55c88279f