Privacy: remove org ID prop from Message
With the decision to separate the new privacy world from the old BC
world, we don't need an org ID override on Message, since we'll either
always infer it from the message/thread (new world) or from the org ID
in the token (old world). Bascially this component can't create a new
thread so we don't need an org ID.

This is technically a BC break, but this param was undocumented and
didn't do anything unless you passed a feature flag anyway.


Test Plan: TS is happy.

Reviewers: yeo-yeo, mhtien, flooey

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/6789

monorepo-commit: 3d52f1278fe4eef826f09a485fd3699b602eb610