Privacy: remove org ID filter from presence reads
Per discussion last week, filtering "presence records" down by their org
ID is confusing. It's not quite sure *how* an org ID filter on presence
should work (in a way that's not subtly/dangerously inconsistent with
the rest of our API). So for now just removing that functionality and we
can come back to it later -- you can always filter down the presence
results you get by hand if it's a huge blocker for someone (as
ugly/painful as that would be to do in some cases).

This keeps the org ID available when *setting* presence, which is a lot
more understandable (and is a primitive you can't simulate otherwise).

This is technically a BC break, but this param was undocumented and
didn't do anything unless you passed a feature flag anyway.


Test Plan: TS is happy.

Reviewers: yeo-yeo, flooey, mhtien

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/6790

monorepo-commit: 2707a2a0092acaa48f394ec461c2016dcd7072e5