Privacy: rename "organization ID" to "group ID" in unreleased props
Per discussion last week, "group ID" is a much preferrable term,
especially in the context of the new privacy API. We'll need to do a bit
of work for that rename in a backwards-compatible way -- but it's
actually really easy for a few spots in our API we added specifically
for the privacy work, since they are undocumented and don't do anything
unless you pass a feature flag anyway. So we can just ignore BC for
those and wholesale rename them.


Test Plan:
TS is happy. Can render a thread from another org in the
testbed.

Reviewers: mhtien, yeo-yeo, flooey

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/6791

monorepo-commit: c41109d8a2ec43812921d4f432907850cdcd0284