Add resolved status filter to threadCounts api
In the previous diff we created the new API and got the metadata and other
filters all functional.
This diff sets up filtering on `resolvedStatus` as well to match up
the filters we have on all our other thread components and APIs.


Reviewers: flooey

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/6597

monorepo-commit: 6d30c1536869e75836c145013c06392af4dcf7b2