Add a field for empty threads in threadsActivitySummary query
I noticed that the `total` value returned in locationSummary(renamed to
threadCounts) does not account for empty threads or threads in which all
messages have been deleted.

Instead of changing the logic, it makes more sense to be more explicit in
our docs and add a new field for empty threads instead.


Reviewers: elenaperezrioja, flooey

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/6631

monorepo-commit: b22c52453cd72e6687c606b7cbe991ee8bd81365