Web components: Change attributes from boolean to string
Boolean isn't a valid data attribute value, it can only be a string.
This was never a problem, until React version `18.3.0-experimental-09285d5a7-20230925`.
This changes the behaviour as such: if you pass `false` to a data attribute of a Web Component, it will be removed from the DOM.
Thus, `?? false` means no data attribute on the DOM, which breaks our `(this.getAttribute('use-shadow-root') === 'false' ? false : true);` logic. (Which previously used to work because the data attribute would be there, and it would have `="false"`)


Test Plan:
In Cordoku, modify  the `node_modules`'s `@cord-sdk/react` mjs and cjs `index` to confirm the new React behaviour described above. Doing this because only Cordoku is getting the new experimental version of React, so we couldn't reproduce anywhere else.
I.e. check that passing `false` results in no data attribute (and hence shadowRoot thread), and passing `"false"` results in a data attribute (and hence no shadowRoot, because we check for `this.getAttribute`)

Load Docs locally, check `buffer-events` is still there and it's `false` when the component loaded correctly.

Reviewers: lazybean, flooey

Reviewed By: lazybean

Pull Request: https://github.com/getcord/monorepo/pull/6879

monorepo-commit: 95cb515c8f9c769355d9e1ac50a95bbdc385489d