Refactoring how ThreadedComments check the cordSDK init
This PR adds the CordContext so we get called back when the cordSDK is initialized.
Previously, we directly checked the window.CordSDK object. This happened to work, as the
component re-renders when we fetch the threads from our API. Instead of having this work as
a side-effect, I opted to use the CordContext so the component re-renders directly as a
consequence of the sdk being initialized.

Test Plan:
Threaded comments still renders correctly in the testbed and the docs. There are
no changes. The main error that would occur if the sdk was not initialized would be the
translations framework showing the placeholders instantaneously before actually rendering the
text, which was the reason this check was added in the first place (#6536). This does not happen,
as seen in the Open/Resolved tab labels and the "load more" button label.


https://github.com/getcord/monorepo/assets/102963329/1b2b135f-6520-4110-be0d-d9623ff87c4b





Reviewers: flooey, jwatzman

Reviewed By: jwatzman

Pull Request: https://github.com/getcord/monorepo/pull/6968

monorepo-commit: c0649635ca789284850302a331f6ea0d1a2d769a