Refactoring the ThreadList implementation in ThreadedComments
The fetching and rendering of threads will have to be reused in #6963 in order to render resolved threads in the same page as open threads in ThreadedComments. This PR extracts all the relevant logic into a separate component to make this change cleaner.

This PR does not contain any functional changes, it just moves the call to the thread API and the rendering of the resulting threads within a new child component called `ThreadList`. It also passes all relevant props and callbacks into the new component.

Test Plan:
ThreadedComments still work as expected. Also tested in #6963
+ I have tested the onRender event and made sure it only fires the first time we render threads.

https://github.com/getcord/monorepo/assets/102963329/d0728357-899b-4a65-add7-894bf8281178

Also switching between open/resolved tabs:


https://github.com/getcord/monorepo/assets/102963329/5d3cc2c9-bf74-4308-9dd8-d250ca96f9ca





Reviewers: jwatzman, flooey

Reviewed By: jwatzman

Pull Request: https://github.com/getcord/monorepo/pull/6970

monorepo-commit: 3417548dd7179a76583b87c85385dae2547281c8