ThreadedComments: Showing resolved threads in the same way as ThreadList
A customer requested that we show resolved threads in ThreadedComments in the same way we do in ThreadList. That is, open threads first, then a button that expands resolved threads, then the resolved threads. All in the same page (Also refer to the test plan for an example). I have added an option to the `displayResolved` property called `sequentially`. LMK if you have other naming suggestions.

The main issue that made this harder to implement is that the thread API is paginated, so we have to separately do fetches of open and resolved threads. To achieve this, I abstracted the list of threads of the ThreadedComments into a separate component called `ThreadList` in #6970.

Note: The order of open threads -> expandResolvedButton -> resolved threads, depends directly on the message order. So the aforementioned order applies when the newest thread is on top. In the opposite case, we completely reverse the order of these components. To do this cleanly, I added these [three components in an array](https://github.com/getcord/monorepo/pull/6963/files#diff-f28a52253835f6ba07b6aaa0aba09e12629e2494296b26cfe52f7e277562c8d5R278), which is [reversed](https://github.com/getcord/monorepo/pull/6963/files#diff-f28a52253835f6ba07b6aaa0aba09e12629e2494296b26cfe52f7e277562c8d5R304) based on the message order. Not sure if this is less readable - it looked the nicest to me.

Test Plan:
(Please excuse my beautiful headshots in this test plan, they are leftovers from testing an image attachment issue 🙂)
Note: There is a bit of a pause in the middle of the video, it is just me switching the message sort order.

https://github.com/getcord/monorepo/assets/102963329/41325466-1190-4c87-ab35-6b8739ff883c


Reviewers: flooey, AlbertQM, dmmiller, jwatzman

Reviewed By: jwatzman

Pull Request: https://github.com/getcord/monorepo/pull/6963

monorepo-commit: b9ed282ee638026389d20d0e111a21f710612fb2