Adding sorting by the thread newest replied to in ThreadedComments
A customer requested to sort threads by the most recent message timestamp. In the process of doing so, we also decided to modify the ThreadedComments props, to be closer to the underlying API parameters, as we spent a bunch of time figuring out what the better API is over the past few months.

The underlying API takes the `sortBy` and `sortDirection` parameters. The `sortBy` is reasonable to expose in ThreadedComments, as it can take the values of `first_message_timestamp` or `most_recent_message_timestamp`, which are useful for ThreadedComments. On the other hand, `sortDirection` can either be `ascending` or `descending`, meaning that the oldest or newest message respectively will be returned first. In the context of ThreadedComments, it doesn't make sense to return the oldest message first (at least for now), as typing a new message in the composer would just disappear (because it would be the newest one). So in the context of ThreadedComments,`sortDirection` should always be `descending`.

Deciding which way threads were displayed though is still important. As mentioned before, we always want to fetch threads starting from the newest one, but we may want to show the newest one on the top, then scrolling down to older ones, or the opposite. So we decided to deprecate the `messageOrder` property and add  `scrollDirection` property instead, which can take the values of `up` or `down` (Open to naming suggestions). So developers can now specify a combination of `sortBy` and `scrollDirection` to achieve the sorting they are looking for.

This PR marks the `messageOrder` as deprecated, adds the two new `sortBy` and `scrollDirection` properties to ThreadedComments and includes the relevant docs and changelog. I have intentionally not given a default value to `scrollDirection` so we can check if developers have actually set it, then give it precedence, otherwise just convert the existing `messageOrder` prop to a `scrollDirection`.

Test Plan:
Testing the `sortBy="most_recent_message_timestamp"` with a button:

https://github.com/getcord/monorepo/assets/102963329/3308ed47-5435-4dca-a39e-f64018a2aeaa

Testing both properties on the docs page:


https://github.com/getcord/monorepo/assets/102963329/86781448-73b7-4369-b835-a42533688396



Reviewers: dmmiller, jwatzman, flooey

Reviewed By: jwatzman

Pull Request: https://github.com/getcord/monorepo/pull/6988

monorepo-commit: b66a31dafcf9115e8b3451b2bcaacaf1ac05da65