Remove positional message ID in sendMessage (#6949)
The `sendMessage` API used to take a message ID in the second
position, but we moved that into the data object, which is a little
easier to use (and allows you to omit it, which is nice).
go/deprecations shows nobody is using the positional argument, so
remove support for calling it that way.

While we're at it, remove the ID from the code example for
`sendMessage`, which is a little nicer.

Test Plan:

Running the example code from the docs still works

monorepo-commit: de26dc9d35d95be946b4d4ab5581fd3a89ab9ed4