Add the ability to set a typing user to the JS API
Add this as a boolean on UpdateThread which indicates that the current user is typing.  This will then just flow through our system as if the user is typing.


Test Plan:
Using testbed with two users

On Andrei
window.CordSDK.thread.updateThread("abc123", {typing:true})

I can then see on Nimrod in that Thread that Andrei is typing

Reviewers: flooey

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/7071

monorepo-commit: fe666329b515827acbf96ea6f44ca971a8a80b82