Implement translationKey in API (#7088)
Allow reading and writing translationKey in our API.  This is a lot of
changes but effectively all of it is plumbing through values.

Part of E-2882

Test Plan:

Sent a new message through the REST API with a `translationKey` set,
it appeared with the correct value when read from REST API and JS API.

Did the same through the JS API, it appeared correctly.

Set a `translationKey` on a message in a `beforeMessageCreate` hook,
it appeared correctly.

monorepo-commit: 99c81e2546943c0934f88c864063d9ad8b25bc13