Translate action messages in UI (#7089)
Add support for handling `translationKey` in the components.  This is
done by trying to look up a translation in the `message_templates`
namespace with that name.  Ours are prefixed with "cord" to keep the
namespace clean for any messages or structures that customers might
want to add.

This takes advantage of i18next's ability to generate objects as
translations, rather than just strings, to produce a correct message
object for the translation.  I decided to have it support returning a
string as well, in which case it will just produce a message that
contains only that string and display it.

In order to allow mentions to be moved around, I pass those into the
message in replacements named `mention1` through `mentionN`.  i18next
doesn't have a way to interpolate in an entire object, so I had to
extract the two strings and pass them and then the message has to
reassemble them, which is annoying but works.

If no translationKey is set or not translation for that key exists, we
render the message's actual content (which already exists), which
should be a good fallback.

Part of E-2882

Test Plan:

Resolved and unresolved a thread, then added translations for the two
keys that we produce, they were rendered properly.

monorepo-commit: 7e867056f92062bf0ff067c8a1cf869d9ece98bb