Add subscribers to the Create Thread REST API
Add the ability to specify subscribers from the get go when creating a thread via the REST.  This is a list of userids that can be specified.  They will be automatically subscribed to the thread.  This is available when creating the thread directly from `/v1/threads` or when creating a thread as part of a message at `/v1/threads/<thread_id>/message`


Test Plan:
Create a thread via the REST API, specify users as subscribers
Look in the db and verify that they are listed there

Reviewers: flooey, jwatzman, km-nur, elenaperezrioja

Reviewed By: elenaperezrioja, km-nur

Pull Request: https://github.com/getcord/monorepo/pull/7191

monorepo-commit: e29b72eddad1a372396a9b2330c61cd3f63ace5c