Correctly setting the flex of the ThreadedComments resolved tabs
When using the `displayResolved` tabs in the canvas demo apps ThreadedComments, I noticed that the "Open" and "Resolved" buttons were not of equal width. Digging into the issue, I noticed that the initial width of the containers was based on the width of the content of each tab. The difference of ~22px difference between tabs was not evident in larger ThreadedComments components, due to proportions, but when the component width was squeezed, it became evident.

This PR correctly sets the flexBasis CSS property on the `displayResolved` tabs to make sure they are always of equal width.

Test Plan:
The before and after videos show the `displayResolved` tabs of ThreadedComments in the canvas app. In both videos, I am hovering the cursor over each tab to show their width while I am increasing the total width of the ThreadedComments component. Before this change, the width of the "Resolved" tab consistently stays ~22px larger than that of the the "Open" tab. After this change, the width of the two tabs is always equal.

Before:


https://github.com/getcord/monorepo/assets/102963329/8e2c20b9-2f6b-4903-91e2-e8c64e14c541


After:

https://github.com/getcord/monorepo/assets/102963329/67a62034-defa-43ec-9b88-c2aa44acd832




Reviewers: jwatzman, flooey

Reviewed By: jwatzman

Pull Request: https://github.com/getcord/monorepo/pull/7227

monorepo-commit: 929987b1bb707824424db9cc3f21aed865075921