More translation strings (#7248)
Translate more strings.  Mostly this is straightforward replacement of
strings with translation items.

One behavior change: in the sidebar, we'll now display "Username
deleted this message" instead of "This message was deleted" in the
message list if the first message of a thread is deleted.  It seemed a
little pointless to have two different messages for the same state,
one for outside the thread and one for inside the thread.

Test Plan:

Switched to cimode, all of the expected places ended up showing their
keys.

monorepo-commit: 5f11b035c92bb28acec79fd196814a3c28e21936