ThreadedComments: fix warning about key on list
PR #6963 introduced the ability to display open and resolved threads sequentially, in one page. It did so by creating two lists of threads, one only with open threads and a button in between them to expand the resolved list. Because of the different sorting options of the ThreadedComments, the order of these 3 components may be reversed, so a decision was made to add them in a list, so that can be done easily.  React uses the "key" value to add/remove/re-sort components in a list, so we need to provide unique keys to each of these components.

PR #7045 added keys to the two ThreadLists, but not to the button in between. This PR adds a key to the button in between to silence the dev console warning.

Test Plan:
The console warning about each child having a unique "key" prop, no longer appears


Reviewers: flooey

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/7282

monorepo-commit: 173219507602e25291ffd0df8b8b276d2eae4ac2