Send and display mouse clicks for LiveCursors
Another attempt to implement clicks.

Add new empty divs component with default styling to represent user clicks.

Add a prop clickDisplayLength to control how long clicks are registered for.

A class name is used for the styling that can be overridden.
Props added to control sending and showing of clicks that default to false.

This approach introduces another hook that takes in the result of the
useUsersCursors hook to computes and store the locations of clicks.


Test Plan: 

Reviewers: flooey, jwatzman, lazybean

Reviewed By: jwatzman, lazybean

Pull Request: https://github.com/getcord/monorepo/pull/7134

monorepo-commit: 1c2beafff572328f3b7761c36efaee00fffec3e7