Live cursors: Adding a null output type to the eventsToLocationFn
While working with the LiveCursors component on the canvas we realized there are cases where the mouse events are not relevant to the component. For example, the case where the canvas has not yet been initialized, but maybe the LiveCursors component has. In that case, the `eventToLocation` function of the LiveCursors did not give us a return type to indicate an invalid location.

This PR enables sending an invalid location via adding a `null` return type to the eventsToLocation function.

Test Plan:
In the LiveCursors docs page, I returned null when the cursor entered the "When to use" section div. Live cursors worked correctly and the presence cleared correctly when entering the div.

https://github.com/getcord/monorepo/assets/102963329/34b528f8-1389-460d-acc1-3b3884cc269f




Reviewers: flooey

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/7286

monorepo-commit: f274ee8fa3869f28a10727e7e4a6f14b1cd692da