Hiding "Show resolved threads" button from ThreadedComments if there are no resolved threads
Working with the ThreadedComments component I noticed that when displaying resolved threads with the `displayResolved="sequentially"` prop, the button to expand the resolved threads section was always present. Given that the label of the button is "Show/Hide resolved threads", it doesn't make sense for it to be present if there are no resolved threads.

This PR checks if there are any resolved threads in the location specified by the ThreadedComments and only renders the button if there are any.

Test Plan:
Tested creating a thread, then resolving it and making sure the "Show resolved threads" button only appears after resolving. Then tested creating an open and resolved thread in a nested location, which worked the same as the previous time. Went back to the original location and enabled partial matching. Two open and one resolved thread appeared. The original open one and then one open and one resolved from the nested location. The "Show resolved threads" button appeared after enabling partial matching, as it counted the resolved thread from the nested location.
(Video same as test plan description)


https://github.com/getcord/monorepo/assets/102963329/64bccafd-1109-43f0-9398-687639c0b773





Reviewers: flooey

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/7301

monorepo-commit: 0a0bf90e361f16df587e272156e4a86287731e1d