Allowing a separate empty state for the Open only and Resolved only sections of ThreadedComments
ThreadedComments can display threads all in one page or in tabs. In the cases where we are showing tabs, we have two empty states, one for the"Open" and one for the "Resolved" threads tab. The existing empty state text is not that relevant to the tab only displaying resolved threads.

This PR adds separate text when displaying resolved threads only and makes sure that it is also customizable separately via the translations framework.

Test Plan:
* Uses of the EmptyStateWithFacepile component were in 1) EmptyThread and 2) ThreadedComments components.

1) The EmptyThread is used in the 'inline' and 'fullHeight' thread, which is used in the Thread, Sidebar, FloatingThread and Inbox components. The empty states of the former two were tested and displayed the correct placeholder text. The latter two cannot have empty threads.

2) Tested all variations of the `displayResolved` prop in ThreadedComments with no threads inside. All instances showed the existing placeholder, except for the "Resolved" tab in the "tabbed" option and the main view in the "resolvedOnly" option, which showed the resolved placeholder.



Reviewers: flooey

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/7266

monorepo-commit: e8b5d4b8f6a6b9167c964e52084874fdcc032188