ThreadedComments scrollbar styling
Because MacOS removes scrollbars when a user is not actively scrolling there is no proper margin between the content of a div and the scrollbar. That causes the scrollbar to overlap with the options menu and reactions buttons, which makes them hard to click. Also, default scrollbars in ThreadedComments on MacOS do not look good in terms of styling.

This PR fixes these issues by matching the scrollbar styles of the google docs ThreadedComments-like component (see image below).

<img width="269" alt="image" src="https://github.com/getcord/monorepo/assets/102963329/10e7de05-07e7-476e-a676-90e39754a596">

Note: On Firefox, we could not figure out a way to make the scrollbar not appear on top of the options menu and reactions. Given that Firefox + MacOS users are a small subset of our users, we decided not to invest more time or create a complicated solution for this. (Though the rest of the styling has been added)

Test Plan:
Before this change, the scrollbar overlaps with the options menu and reactions buttons and disappears when we are not scrolling. After this change, the scrollbar is always present, and no longer overlaps with the buttons (except on Firefox).

Before:

https://github.com/getcord/monorepo/assets/102963329/f584fe04-53a6-44a1-9503-6f5853b579fd

After:

| OS / Browser      | Chrome | Safari | Firefox | Edge |
| ---------------- | -------- | ------ | ------- | ----- |
| MacOS (default scrollbars) | ok | ok | scrollbar overlaps with options menu (as expected) | ok |
| MacOS (scrollbars always on) | ok | ok | ok | ok |

I could not figure out a way to check on Windows. I will be checking in staging, after landing this.

In all test cases (except Firefox on MacOS default scrollbars), there is a ThreadedComments component with no overflow on the left and one with overflow on the right. On the one on the left, the options menu should only have an 8px padding from the edge of its container, whereas the one on the right should have an 8px padding plus the scrollbar to its right.

MacOS scrollbars (default):

Chrome:

https://github.com/getcord/monorepo/assets/102963329/60908a2f-3995-4540-bd96-dc28fadc8c79


Safari:

https://github.com/getcord/monorepo/assets/102963329/02c6cb8c-0862-4d38-b833-226989a22764


Firefox :

https://github.com/getcord/monorepo/assets/102963329/2b2c5087-5d3d-4673-a9bf-f638fd7eaeb9


Edge:

https://github.com/getcord/monorepo/assets/102963329/07d624a9-200f-4057-9812-6ddc15602e47



MacOS scrollbars (Always on):

Chrome:


https://github.com/getcord/monorepo/assets/102963329/55e7b3fd-3de7-42ad-b236-c47c7a489501


Safari:

https://github.com/getcord/monorepo/assets/102963329/14b4f0ca-d837-4e8b-af2b-d1b75d5957b3


Firefox:

https://github.com/getcord/monorepo/assets/102963329/bb554bc7-723c-400f-97e4-3fd0cbe3fd94


Edge:


https://github.com/getcord/monorepo/assets/102963329/ad6fd2d9-87dd-4bd2-84fd-7c51b31f48ff




Reviewers: AlbertQM, lazybean

Reviewed By: lazybean

Pull Request: https://github.com/getcord/monorepo/pull/7155

monorepo-commit: 98403fe767f6e6f536f0933450736db8d1197da6